src/backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-filter
PyJWT
pytz
sqlparse
yapf
psycopg2-binary
python-dotenv
whitenoise
gunicorn
whitenoise[brotli]
Faker
pillow
uuid
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
django-otp
pyotp
qrcode[pil]
twilio